feat: 主键更新

+++ mcdull-bussiness/mcdull-admin/src/main/java/io/gitee/dqcer/mcdull/admin/model/entity/sys/LogDO.java
@@ -21,12 +21,12 @@ public class LogDO extends RelDO {
    /**
     * 操作人的账号主键
     */
-    private Long accountId;
+    private Integer accountId;
 
    /**
     * 租户主键
     */
-    private Long tenantId;
+    private Integer tenantId;
 
    /**
     * 客户端ip
