+++ mcdull-framework/mcdull-framework-base/src/main/java/io/gitee/dqcer/mcdull/framework/base/storage/UserContextHolder.java
@@ -57,7 +57,7 @@ public class UserContextHolder {
      *
      * @return {@link Long}
      */
-    public static Long currentUserId() {
+    public static Integer currentUserId() {
         return UNIFY_SESSION.get().getUserId();
     }
