+++ mcdull-framework/mcdull-framework-starters/mcdull-framework-starter-security/src/main/java/io/gitee/dqcer/mcdull/framework/security/AbstractUserDetailsService.java
@@ -17,14 +17,14 @@ public abstract class AbstractUserDetailsService implements StpInterface {
     public List<String> getPermissionList(Object loginId, String loginType) {
         // 返回此 loginId 拥有的权限列表
         // TODO: 2023/12/1 缓存
-        return permissionList(Convert.toLong(loginId));
+        return permissionList(Convert.toInt(loginId));
     }
 
     @Override
     public List<String> getRoleList(Object loginId, String loginType) {
         // 返回此 loginId 拥有的角色列表
         // TODO: 2023/12/1 缓存
-        return this.roleList(Convert.toLong(loginId));
+        return this.roleList(Convert.toInt(loginId));
     }
 
     /**
@@ -33,7 +33,7 @@ public abstract class AbstractUserDetailsService implements StpInterface {
      * @param loginId 登录id
      * @return {@link List}<{@link String}>
      */
-    protected abstract List<String> permissionList(Long loginId);
+    protected abstract List<String> permissionList(Integer loginId);
 
     /**
      * 角色列表
@@ -41,5 +41,5 @@ public abstract class AbstractUserDetailsService implements StpInterface {
      * @param loginId 登录id
      * @return {@link List}<{@link String}>
      */
-    protected abstract List<String> roleList(Long loginId);
+    protected abstract List<String> roleList(Integer loginId);
 }
