+++ mcdull-framework/mcdull-framework-starters/mcdull-framework-starter-web/src/main/java/io/gitee/dqcer/mcdull/framework/web/interceptor/BaseInfoInterceptor.java
@@ -99,7 +99,7 @@ public abstract class BaseInfoInterceptor implements HandlerInterceptor {
                 return false;
             }
             UserSession userSession = result.getData();
-            Long userId = userSession.getUserId();
+            Integer userId = userSession.getUserId();
             unifySession.setUserType(userSession.getType());
             unifySession.setUserId(userId);
         }
