+++ mcdull-support/mcdull-uac/mcdull-uac-provider/src/main/java/io/gitee/dqcer/mcdull/uac/provider/model/vo/RoleVO.java
@@ -18,7 +18,7 @@ public class RoleVO implements VO {
     /**
      * 主键
      */
-    private Long id;
+    private Integer id;
 
     /**
      * 创建时间
@@ -28,7 +28,7 @@ public class RoleVO implements VO {
     /**
      * 创建人
      */
-    private Long createdBy;
+    private Integer createdBy;
 
     private String createdByStr;
 
@@ -40,7 +40,7 @@ public class RoleVO implements VO {
     /**
      * 更新人
      */
-    private Long updatedBy;
+    private Integer updatedBy;
 
     private String updatedByStr;
 
@@ -122,11 +122,11 @@ public class RoleVO implements VO {
     }
 
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -138,11 +138,11 @@ public class RoleVO implements VO {
         this.createdTime = createdTime;
     }
 
-    public Long getCreatedBy() {
+    public Integer getCreatedBy() {
         return createdBy;
     }
 
-    public void setCreatedBy(Long createdBy) {
+    public void setCreatedBy(Integer createdBy) {
         this.createdBy = createdBy;
     }
 
@@ -162,11 +162,11 @@ public class RoleVO implements VO {
         this.updatedTime = updatedTime;
     }
 
-    public Long getUpdatedBy() {
+    public Integer getUpdatedBy() {
         return updatedBy;
     }
 
-    public void setUpdatedBy(Long updatedBy) {
+    public void setUpdatedBy(Integer updatedBy) {
         this.updatedBy = updatedBy;
     }
