+++ mcdull-support/mcdull-uac/mcdull-uac-provider/src/main/java/io/gitee/dqcer/mcdull/uac/provider/web/dao/repository/IMenuRepository.java
@@ -21,7 +21,7 @@ public interface IMenuRepository extends IService<MenuDO> {
 
     List<String> allCodeList();
 
-    Map<Long, List<String>> menuCodeListMap(Map<Long, List<Long>> menuListMap);
+    Map<Integer, List<String>> menuCodeListMap(Map<Integer, List<Integer>> menuListMap);
 
-    List<MenuDO> list(Collection<Long> collection);
+    List<MenuDO> list(Collection<Integer> collection);
 }
